requirements.txt
requests>=2.31
pydantic>=2.6
PyYAML>=6.0.1
rapidfuzz>=3.5
jinja2>=3.1
feedparser>=6.0
sentence-transformers>=2.2
faiss-cpu>=1.7
numpy>=1.24
python-dotenv
python-dotenv>=1.0
